requirements.txt
pyssc
rpi_lcd
RPi.GPIO